[feat] : Label 없으면 재 생성 및 전체 예외 처리

+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/label/domain/LabelRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 public interface LabelRepository extends JpaRepository<Label, Long> {
     List<Label> findByApplicantId(String applicantId);
 
-    Optional<Label> findByApplicantIdAndIdpId(String applicantId, Integer idpId);
+    Optional<Label> findByApplicantIdAndIdpId(String applicantId, Long idpId);
 
     List<Label> findByCardIdIn(List<Long> cardIds);
 }
